swtpm: Add compiler attribute for printf to logging functions
Add the printf compiler attribute to the logging functions and fix detected
parameter issues.

+++ src/swtpm/tlv.c
@@ -93,7 +93,8 @@ tlv_data_append(unsigned char **buffer, uint32_t *buffer_len,
 
     tmp = realloc(*buffer, (size_t)totlen);
     if (!tmp) {
-         logprintf(STDERR_FILENO, "Could not allocate %u bytes.\n", totlen);
+         logprintf(STDERR_FILENO, "Could not allocate %zu bytes.\n",
+                   (size_t)totlen);
          return TPM_FAIL;
     }
     *buffer = tmp;
